separate index.js into modules
- addNew for adding a new book
- popWin for informing a new book was successfully added
- removeBook for deleting a book
- showIt for toggling navigation
- showBooks for listing all books
- showNav for showing navigation
- showTime for show time
- toggleActive just for highlighting current navigation page.
