+++ index.html
@@ -6,8 +6,13 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-    <link rel="stylesheet" href="style.css">
-    <script src="./index.js"></script>
+    <link rel="stylesheet" href="index.css">
+    <script src="./modules/showIt.js"></script>
+    <script src="./modules/toggleActive.js"></script>
+    <script src="./modules/removeBook.js"></script>
+
+
+    <script type="module" src="./index.js"></script>
 
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
