+++ modules/removeBook.js
@@ -1,0 +1,8 @@
+const removeBook = (id) => {
+  let addBooks = JSON.parse(localStorage.getItem('addBooks'));
+  addBooks = addBooks.filter((e) => e.id !== id);
+
+  localStorage.setItem('addBooks', JSON.stringify(addBooks));
+};
+
+removeBook();
